LTT: Remove temporary/internal fields from saved settings
Some temporary fields such as "curentSpeed" were accidentally saved in the laser settings file. This could potentially cause bugs, as these variables should always be properly initialized and not remember the state from the previous execution of VisiCut.

Please note that this breaks compatibility with existing settings. You will have to recreate the lasercutter in VisiCut. However, as far as I know this only affects "my" lab, as nobody else is using this driver yet.