Epilog: Raise warning message if starting point is unsupported
The Epilog driver code currently only supports starting points top-left of the job's bounding box.
Otherwise, our code generates data which is at least partly wrong and is different from what the manufacturer does.

Because we couldn't agree on making this a hard error, it is a warning message until we have more user reports.