Add RasterBuilder (and dependencies).
Depends on RasterElement and AbstractLaserProperty. Requires slightly reduced declarations on VectorCommand. 

RasterBuilder is works as an iterable/iterator ( you can use it in a for-loop) and is based on easily debuggable states as you would find in a state machine. It requires and allocates nearly no memory. It reuses the VectorCommand interface so already existing Vector code can be easily reused. Permits either axis rastering, starting in any of the four corners.