Join broken paths on inner-first sorting with tolerance
This is helpful for files from DXF with polylines that don't exactly match up.

The old approach only worked for exactly identical start/end points, which is problematic because,
e.g., in floating point ((a + b) + c) == ((a + b) + c) isn't always true.
Now we use a tolerance of 0.9 pixels (in the current DPI), so about 0.02mm for 1000dpi.

INNER_FIRST sorting is now a little slower than before, but still a lot faster than NEAREST.
A worst-case example file with 80.000 paths before joining and 1500 after takes about 3 seconds to process (and 50 seconds for NEAREST).
This time grows roughly linear if the paths start/end at points spaced roughly even in X. In the worst case, if all paths end at the same X, it grows quadratically.