Ruida: force first movement
When the machine code is generated, the initial position of the laser
is unknown. The initial state of the hardware might differ from [0,0].

Even worse: the instantiation of the driver is reused across multiple
laser runs, so the driver retains the position of the last movement, but
the controller might move to a different position after execution has
finished.

Instead of assuming a laser position, always force the generation of
machine code for the first movement by setting last_xy to NaN and
adding code to handle moves from an unknown position.

Fixes: #189