NearestVectorOptimizer: Connect broken paths
For files where a continuous path was not one path, but multiple separate line objects, the lasercutter movement was very jerky when using NEAREST mapping in VisiCut.
This is now solved: VisiCut detects broken paths and re-joins them (up to a small tolerance of almost 1 pixel).

Note that the INNER_FIRST sorting was not affected by the problem. There, the same solution was already used.